pyproject.toml
[project]
name = "shadow-stories"
version = "0.2.0"
description = "Live shadow puppet narrator for ages 4–8, powered by Gemini"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.64.0",
    "python-dotenv>=1.0",
    "websockets>=12.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "sympy>=1.13,<1.14",
    "httpx>=0.28.1",
]

[project.scripts]
shadow-narrate = "shadow_stories.cli.__main__:cli_main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff",
    "mypy",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.setuptools.packages.find]
include = ["shadow_stories*"]
exclude = ["models*"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.2",
]
